refactor(engine): split DrawBatcher into batches/ sub-modules (T9-split)
Pure mechanical file reorganization — zero behavior/signature/visibility change.
batches.rs had reached 1480 LOC against the C1 <1500 per-file cap; the T9e
image/texture family (~770 LOC) cannot land without first splitting the single
`impl DrawBatcher` across files.

batches.rs -> batches/:
  - mod.rs (238)       struct + new + core helpers (add_tessellated_with_key,
                       submit_transformed_geometry, finish_current_segment,
                       prime_tessellator_scale, shader_to_gradient_stops)
  - shapes.rs (571)    rect/rrect/circle/oval/draw_drrect/draw_arc
  - gradients.rs (354) gradient_rect/radial/sweep/shadow_rect/dispatch_shader_rect
  - paths.rs (386)     draw_path/draw_vertices/line/draw_shadow

All 21 methods relocated byte-identical. The only token changes are required
consequences of the file->dir split: per-family `impl` methods move from
`pub(super)` to `pub(in super::super)` (identical effective scope = the `wgpu`
module) and `super::` imports deepen to `super::super::`. Each `impl DrawBatcher`
block carries its own clippy cast allows. painter.rs and wgpu/mod.rs untouched.

C1: each file is now well under 1500 (largest 571), leaving room for T9e's
batches/images.rs. Gates: fmt/taplo/clippy(both modes)/doc/typos/port-check
green; GPU-readback serial (DX12) 210/0/7 byte-identical. rust-reviewer verified
pure move (21/21 bodies identical, visibility equivalent, zero collateral).
